finish fixing breaking changes from bumping `@web5/*` packages
