+++ packages/protocol/src/main.ts
@@ -18,7 +18,6 @@ Resource.factory = resourceFactory
 export * from './resource-kinds/index.js'
 export * from './message-kinds/index.js'
 export * from './did-resolver.js'
-export * from './dev-tools.js'
 export * from './crypto.js'
 export * from './types.js'
 export { Message, Resource }
